Autoload & Build

+++ build.php
@@ -6,16 +6,29 @@ if (php_sapi_name() !== 'cli') {
 }
 
 // Any command needed to run and build plugin assets when newly cheched out of repo.
-$buildCommands = [
-    'npm ci --no-progress --no-audit',
-    'npm update @helsingborg-stad/styleguide',
-    'npx --yes browserslist@latest --update-db',
-    'npm run build'
-];
+$buildCommands = [];
 
 //Add composer build, if flag --no-composer is undefined.
-if(is_array($argv) && !in_array('--no-composer', $argv)) {
-    $buildCommands[] = 'composer install --prefer-dist --no-progress --no-dev'; 
+//Dump autloader. 
+//Only if composer.json exists.
+if(file_exists('composer.json')) {
+    if(is_array($argv) && !in_array('--no-composer', $argv)) {
+        $buildCommands[] = 'composer install --prefer-dist --no-progress --no-dev'; 
+    }
+    $buildCommands[] = 'composer dump-autoload';
+}
+
+//Run npm if package.json is found
+if(file_exists('package.json') && file_exists('package-lock.json')) {
+    $buildCommands[] = 'npm ci --no-progress --no-audit';
+} elseif(file_exists('package.json') && !file_exists('package-lock.json')) {
+    $buildCommands[] = 'npm install --no-progress --no-audit';
+}
+
+//Run build if package-lock.json is found
+if(file_exists('package-lock.json')) {
+    $buildCommands[] = 'npx --yes browserslist@latest --update-db';
+    $buildCommands[] = 'npm run build';
 }
 
 // Files and directories not suitable for prod to be removed.
@@ -26,10 +39,13 @@ $removables = [
     'build.php',
     '.npmrc',
     'composer.json',
+    'composer.lock',
     'env-example',
     'webpack.config.js',
     'package-lock.json',
-    'package.json'
+    'package.json',
+    'phpunit.xml.dist',
+    'README.md'
 ];
 
 $dirName = basename(dirname(__FILE__));
